requirements.txt
discord.py[voice] @ git+https://github.com/imayhaveborkedit/discord.py.git
discord-ext-voice-recv
SpeechRecognition
